+++ pkg/receivers/main.go
@@ -13,17 +13,3 @@ func SetUpReceivers(cond *conductor.Conductor, bus giga.MessageBus, conf giga.Co
 	// Set up configured Callbacks
 	SetupCallbacks(cond, bus, conf)
 }
-
-func StartServices(cond *conductor.Conductor, bus giga.MessageBus, conf giga.Config, store giga.Store) {
-	// BalanceKeeper sends "Balance Change" events.
-	keeper := NewBalanceKeeper(store)
-	cond.Service("NewBalanceKeeper", keeper)
-
-	// InvoiceStamper sends "Invoice Paid" and "Invoice Partial Payment" events.
-	stamper := NewInvoiceStamper()
-	cond.Service("InvoiceStamper", stamper)
-
-	// PayMaster sends "Payment Accepted" and "Payment Confirmed" events.
-	master := NewPayMaster()
-	cond.Service("PayMaster", master)
-}
