+++ pkg/services/balancekeeper.go
@@ -1,4 +1,4 @@
-package receivers
+package services
 
 import (
 	"context"
