+++ pkg/services/paymaster.go
@@ -1,4 +1,4 @@
-package receivers
+package services
 
 import (
 	"context"
